refactor to fix circular dependencies
